fix: prevent integer overflow in watering chart time range calculation
Fixed integer overflow bug that prevented watering history chart from displaying.
The calculation `selectedRange.daysBack * 24 * 60 * 60 * 1000` was performed
with Int arithmetic, causing overflow for ranges >= 25 days:
- 30 days overflowed to -1,702,967,296
- This set rangeStartMs to a future time, filtering out all logs

Changed to use Long arithmetic by converting daysBack with .toLong() to ensure
the time calculation uses 64-bit values instead of 32-bit.

Tests verified: all 8 WateringHistoryChartTest cases still pass (they use small
test time values that don't trigger overflow). Full unit test and lint checks pass.

+++ app/src/main/kotlin/com/yapt/planttracker/ui/components/WateringHistoryChart.kt
@@ -47,7 +47,7 @@ fun WateringHistoryChart(
         .sortedBy { it.loggedAt }
 
     val now = System.currentTimeMillis()
-    val rangeStartMs = now - (selectedRange.daysBack * 24 * 60 * 60 * 1000)
+    val rangeStartMs = now - (selectedRange.daysBack.toLong() * 24 * 60 * 60 * 1000)
 
     val intervals = computeWateringIntervals(wateringLogs, rangeStartMs, now)
